+++ src/sector/mod.rs
@@ -2,6 +2,6 @@
 pub struct Sector {
     /// `altitude >= 0` solid \
     /// `altitude < 0` water
-    pub altitude: f64,
+    pub altitude: f32,
     pub count_of_cells: usize,
 }
